Large system refactoring (addresses memory issues with ASan and simplifies the API). (#189)
## Major changes

### Refactoring

- Change the functionality of `Virus` and `Tool` functions: no longer capturing parameters via pointers (unsafe). Instead, switched to capturing parameter names via `std::shared_ptr` (validity checked in run-time).
- Replaced `GET_MODEL()` (`dynamic_cast`) macros with proper implementation of using `static_cast` for more efficient and easier to maintain models.
- The events are now part of `Model`. The event callback functions were passed via `std::function` to the `Event` class, and later dispatched by `Model`. There was no reason why to keep this design, so, instead, the update functions are called by enum (so instead of copying the function to the `Event` object, only an enum is passed), and then triggered as a private member of `Model`.
- The mixer functions used by the `Model` class are now member functions instead of member objects. Previously, these were stored as `MixerFun` objects, adding overhead and unnecessary dereferencing.
- The `Model::{array_double_tmp, array_virus_tmp}` are now proper arrays (so allocated in the stack) of 1024 x 2.

### Deleted/defunct code:

- Changed the functionality of the `Model::operator()` to not return a reference. This was unsafe behavior. Forcing users to call `Model::set_param()` instead.
- Removed unused (and redundant) constructor method for models. Left a single main constructor in place.
- The method `Agent::rm_agent_by_virus` is now `delete`d. It was not being used, and the method `Agent::rm_virus` already has the same functionality.
- The overhead of `ToolFunctions` that were shared between agents as a smart pointer has been eliminated. The functions are now stored as `ToolFun` objects in `Tool`. The same applies to `VirusFunctions`.

### Bugs

- The `Tool` and `Entity` classes lacked a `clone_ptr()` method. This was causing threads to share entities, which is most likely what triggered the ASan issues.
- One cryptic fix (proposed by Gemini, Codex, and Claude) was to free the memory of `Agent::tools` and `Agent::entities`. This change works with the code base and makes things safer (probably), so we kept it.

---------

+++ examples/13-genint/main.cpp
@@ -42,9 +42,9 @@ int main()
     // Computing the mean of gentime conditioning on gentime >= 0
     double gentime_sir_obs_mean = 0.0;
     size_t n = 0;
-    for (int i = 0; i < gentime_sir_obs.size()*3/4; i++)
+    for (int i = 0; i < static_cast<int>(gentime_sir_obs.size()*3/4); i++)
     {
-        if (time[i] >= max_days_for_calc)
+        if (time[i] >= static_cast<int>(max_days_for_calc))
             continue;
 
         if (gentime_sir_obs[i] >= 0)
@@ -81,9 +81,9 @@ int main()
     // Computing the mean of gentime conditioning on gentime >= 0
     double gentime_seir_obs_mean = 0.0;
     n = 0;
-    for (int i = 0; i < gentime_seir_obs.size()*3/4; i++)
+    for (int i = 0; i < static_cast<int>(gentime_seir_obs.size()*3/4); i++)
     {
-        if (time[i] >= max_days_for_calc)
+        if (time[i] >= static_cast<int>(max_days_for_calc))
             continue;
         if (gentime_seir_obs[i] >= 0)
         {
